+++ src/components/sidebar/_roads.tsx
@@ -1,4 +1,4 @@
-import styles from "@/styles/SubBlocks.module.css";
+import styles from "@/styles/MapParamsCard.module.css";
 import Toggle from "../common/_toggle";
 import { useAtom } from "jotai";
 import { roadsAtom } from "@/atoms";
@@ -7,7 +7,7 @@ import cx from 'classnames'
 export default function Roads() {
   const [roads, setRoads] = useAtom(roadsAtom);
   return (
-    <div className={cx(styles.subBlock, { [styles.active]: roads })}>
+    <div className={cx(styles.card, { [styles.active]: roads })}>
       <h4>
         Show road network <Toggle checked={roads} onChange={setRoads} />
       </h4>
