+++ src/styles/Sidebar.module.css
@@ -3,7 +3,7 @@
   z-index: var(--z-flow-info);
   top: 10px;
   left: 40px;
-  filter: drop-shadow(0px 0px 25px rgba(41, 41, 41, 0.1));
+  filter: var(--shadow);
   width: 368px;
 }
 
@@ -12,7 +12,7 @@
 }
 
 .flowInfo > nav {
-  border-radius: var(--border-radius) var(--border-radius) 0 0;
+  /* border-radius: var(--border-radius) var(--border-radius) 0 0; */
   background: var(--foreground);
   color: var(--background);
   padding: var(--padding);
@@ -51,7 +51,7 @@
 
 .content {
   background: var(--background-80);
-  border-radius: 0 0 var(--border-radius) var(--border-radius);
+  /* border-radius: 0 0 var(--border-radius) var(--border-radius); */
   max-height: 55vh;
   overflow-x: hidden;
   overflow-y: auto;
