+++ src/styles/globals.css
@@ -5,7 +5,8 @@
   --padding-v: 32px;
 
   --foreground-rgb: 0, 0, 0;
-  --background-rgb: 246, 243, 238;
+  /* --background-rgb: 246, 243, 238; */
+  --background-rgb: 255, 255, 255;
   --background-2-rgb: 239, 233, 223;
 
   --foreground: rgb(var(--foreground-rgb));
@@ -31,6 +32,8 @@
   --z-popup-highlighted: 4;
   --z-flow-info: 5;
   --z-search: 5;
+
+  --shadow: drop-shadow(0px 0px 10px rgba(41, 41, 41, 0.1));
 }
 
 * {
